+++ 05 Implementation/New-Yorker designer app/app/src/main/AndroidManifest.xml
@@ -9,9 +9,9 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.NewYorkerUI">
-        <activity android:name=".glasstype"></activity>
-        <activity android:name=".doorhandletype" />
-        <activity android:name=".doortype" />
+        <activity android:name=".Glasstype"></activity>
+        <activity android:name=".Doorhandletype" />
+        <activity android:name=".Doortype" />
         <activity android:name=".Loadingscreen" />
         <activity android:name=".ShowAccept" />
         <activity android:name=".UserContact" />
