+++ 05 Implementation/New-Yorker designer app/app/src/main/java/com/example/newyorkerui/Doorhandletype.java
@@ -1,0 +1,27 @@
+package com.example.newyorkerui;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.widget.Button;
+
+public class Doorhandletype extends AppCompatActivity {
+
+    Button switchButton;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_doorhandletype);
+
+        switchButton = (Button) findViewById(R.id.chooseDoorHandleType);
+        switchButton.setOnClickListener(view -> launchActivity());
+    }
+
+    private void launchActivity() {
+
+        Intent intent = new Intent(this, ShowWall.class);
+        startActivity(intent);
+    }
+}
