+++ 05 Implementation/New-Yorker designer app/app/src/main/java/com/example/newyorkerui/UserMeasurements.java
@@ -21,7 +21,7 @@ public class UserMeasurements extends AppCompatActivity {
         height_input = (EditText) findViewById(R.id.height_input);
         width_input = (EditText) findViewById(R.id.width_input);
 
-        switchButton = (Button) findViewById(R.id.makeoffer);;
+        switchButton = (Button) findViewById(R.id.makeoffer);
         switchButton.setOnClickListener(view -> launchActivity());
     }
 
@@ -30,7 +30,7 @@ public class UserMeasurements extends AppCompatActivity {
         String w = width_input.getText().toString();
         String h = height_input.getText().toString();
 
-        Intent intent = new Intent(this, ShowWall.class);
+        Intent intent = new Intent(this, Glasstype.class);
         intent.putExtra("height", h);
         intent.putExtra("width", w);
         startActivity(intent);
