--- 05 Implementation/New-Yorker designer app/app/src/main/java/com/example/newyorkerui/doortype.java
@@ -1,14 +1,0 @@
-package com.example.newyorkerui;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-public class doortype extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_doortype);
-    }
-}
