+++ 05 Implementation/New-Yorker designer app/app/src/main/res/layout/activity_doortype.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".doortype">
+    tools:context=".Doortype">
 
     <Button
         android:id="@+id/ChooseDoorHandleType"
@@ -108,6 +108,7 @@
         android:layout_marginLeft="100dp"
         android:layout_marginTop="8dp"
         android:text="@string/dubbelt_skyded_r_uden_l_s"
+        app:layout_constraintBottom_toTopOf="@+id/ChooseDoorHandleType"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.0"
         app:layout_constraintStart_toStartOf="parent"
