+++ 05 Implementation/New-Yorker designer app/app/src/main/res/layout/activity_usermeasurements.xml
@@ -9,17 +9,18 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:android="http://schemas.android.com/apk/res/android">
 
-    <Button android:layout_height="wrap_content"
+    <Button
+        android:id="@+id/makeoffer"
         android:layout_width="wrap_content"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintHorizontal_bias="0.498"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:iconPadding="16dp"
-        app:backgroundTint="#BF8E55"
-        android:text="@string/vis_design"
+        android:layout_height="wrap_content"
         android:layout_marginBottom="100dp"
-        android:id="@+id/makeoffer"/>
+        android:text="@string/videre"
+        app:backgroundTint="#BF8E55"
+        app:iconPadding="16dp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.498"
+        app:layout_constraintStart_toStartOf="parent" />
 
     <EditText android:layout_height="wrap_content"
         android:layout_width="wrap_content"
